Bruk Aspect og meterRegistry for å måle tid bruk
