+++ src/main/kotlin/no/nav/melosysskattehendelser/metrics/MeasuredMetricsProvider.kt
@@ -1,0 +1,12 @@
+package no.nav.melosysskattehendelser.metrics
+
+fun interface MeasuredMetricsProvider {
+    fun getMeasured(): List<MethodTimingStatus>
+
+    data class MethodTimingStatus(
+        val method: String,
+        val count: Long,
+        val totalTimeMs: Double,
+        val maxTimeMs: Double
+    )
+}
